added ability to set and get read and write files

+++ rdfapi.php
@@ -29,13 +29,14 @@ SOFTWARE.
 */
 require_once('application.php');
 require_once('ns.php');
-//require_once('/code/arc2/ARC2.php');
 
 class RDFFactory extends Application
 {
 	private $buf = '';
 	private $ns = null;
 	private $types = null;
+	private $read_file=null;
+	private $write_file=null;
 	
 	function __construct()
 	{
@@ -48,12 +49,41 @@ class RDFFactory extends Application
 	function GetRDF() {return $this->buf;}
 	function AddRDF($buf) {$this->buf .= $buf;return TRUE;}
 	function DeleteRDF() {$this->buf = '';return TRUE;}
-	function WriteRDF($fp) {
-		fwrite($fp,$this->buf);
+
+	
+	function SetReadFile($file,$gzcompress=false)
+	{
+		$this->read_file = new FileFactory($file,$gzcompress);
+		return $this->read_file;
+	}
+	function GetReadFile()
+	{	
+		return $this->read_file;
+	}
+	function WriteFileExists()
+	{
+		if(isset($this->write_file)) return TRUE;
+		return FALSE;
+	}
+	function SetWriteFile($file,$gzcompress=false)
+	{
+		$this->write_file = new FileFactory($file,$gzcompress);
+		return $this->write_file;
+	}
+	function GetWriteFile()
+	{
+		return $this->write_file;
+	}
+	function WriteRDFBufferToWriteFile() 
+	{
+		if($this->WriteFileExists() === FALSE) {
+			trigger_error("Write file not set!");
+			return FALSE;
+		} 
+		$this->GetWriteFile()->Write($this->buf);
 		$this->DeleteRDF();
 		return TRUE;
 	}
-	
 	
 	function Quad($s_uri, $p_uri, $o_uri, $g_uri = null)
 	{
